Refactored enumerators into their own package
