+++ app/src/main/java/durm/caloriecounter/enumerators/enumGender.java
@@ -1,4 +1,4 @@
-package durm.caloriecounter;
+package durm.caloriecounter.enumerators;
 
 public enum enumGender {
     MALE,
